+++ salesforce/soap_bodies.py
@@ -412,14 +412,6 @@ retrieve_all_task_body = """
                         </met:types>
                         <met:types>
                             <met:members>*</met:members>
-                            <name>criteriaBasedRules</name>
-                        </met:types>
-                        <met:types>
-                            <met:members>*</met:members>
-                            <name>ownerRules</name>
-                        </met:types>
-                        <met:types>
-                            <met:members>*</met:members>
                             <name>CustomPageWebLink</name>
                         </met:types>
                         <met:types>
@@ -462,10 +454,6 @@ retrieve_all_task_body = """
                         <met:types>
                             <met:members>*</met:members>
                             <name>HomePageLayout</name>
-                        </met:types>
-                        <met:types>
-                            <met:members>*</met:members>
-                            <name>InstalledPackage</name>
                         </met:types>
                         <met:types>
                             <met:members>*</met:members>